✨ feat(frontend): 검색 페이지 추가 및 포트폴리오 수정 기능 개선
- 【feat】 메인 페이지에 검색 버튼 추가, 검색 페이지(`/main/search`) 구현
- 【refactor】 `ArtistPortfolioClient` 및 `ArtistPortfolioEditableClient` 컴포넌트에서 YouTube 썸네일 추출 로직을 별도 컴포넌트(`YoutubeThumbnail`)로 분리
- 【fix】 `dancer_directing` 관계 데이터 선택 시 `directing_id`뿐만 아니라 `title` 정보도 함께 가져오도록 수정, 연출 항목 수정 시 기존 항목 유지/삭제 로직 개선
- 【refactor】 포트폴리오 관련 타입 정의를 별도 파일(`frontend/types/portfolio.ts`)로 분리하여 관리
- 【fix】 프로필 페이지에서 인증 로딩 상태 처리 개선 및 불필요한 로그인 프롬프트 제거
