+++ frontend/components/YoutubeThumbnail.tsx
@@ -1,0 +1,22 @@
+import Image from "next/image";
+
+function extractYouTubeId(url: string): string | null {
+  const regExp = /^.*(youtu.be\/|v\/|u\/\w\/|embed\/|watch\?v=|&v=)([^#&?]*).*/;
+  const match = url.match(regExp);
+  return match && match[2].length === 11 ? match[2] : null;
+}
+
+export default function YouTubeThumbnail({ url, title }: { url: string; title?: string }) {
+  const videoId = extractYouTubeId(url);
+  if (!videoId) return null;
+  return (
+    <div className="relative w-full aspect-video bg-black overflow-hidden">
+      <Image
+        src={`https://img.youtube.com/vi/${videoId}/hqdefault.jpg`}
+        alt={title || 'Video thumbnail'}
+        fill
+        className="object-cover object-center"
+      />
+    </div>
+  );
+}
